requirements.txt
# Core Framework Dependencies
dspy-ai>=2.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Language Model Providers
openai>=1.0.0
anthropic>=0.7.0
cohere>=4.0.0

# OCR and Document Processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
pypdf2>=3.0.0
python-multipart>=0.0.6

# Data Processing and Validation
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Machine Learning and Optimization
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework (for API endpoints)
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic[email]>=2.0.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0

# Logging and Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
httpx>=0.24.0

# Development Tools
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0
mkdocstrings>=0.22.0

# Utilities
tqdm>=4.65.0
rich>=13.5.0
typer>=0.9.0
pathlib2>=2.3.7

# Optional: Cloud Services
boto3>=1.28.0  # AWS services
google-cloud-vision>=3.4.0  # Google Cloud Vision
azure-ai-formrecognizer>=3.2.0  # Azure Document Intelligence

# Optional: Advanced OCR
easyocr>=1.7.0
transformers>=4.30.0
torch>=2.0.0

# Optional: Performance and Caching
joblib>=1.3.0
cachetools>=5.3.0 